Refactor: Update ReadingConfig import path
